2024.05. 28 🛠️ 크루 게시판 작업.. 🛠️ 크루 게시판 디테일 페이지 작업중 ✅ 크루 모임 참여 명단 불러오기 완료 🛠️ 크루 모임 참여 및 미참여에 따른 리스트 추가 작업.

+++ src/main/java/com/lec/sping/domain/crew/CrewBoard.java
@@ -22,7 +22,7 @@ public class CrewBoard {
     private LocalDateTime startDate;// 게시글 시작 일자(공지글, 모임 게시글 해당)
     private LocalDateTime endDate;  // 게시글 종료 일자(공지글, 모임 게시글 해당)
     private LocalDateTime regDate;  // 게시글 생성 날짜
-    private Integer memberCount;    // 참여 인원 설정(모임 게시글 해당)
+    private Long memberCount;    // 참여 인원 설정(모임 게시글 해당)
     private String address;         // 모임 장소
     private String boardType;       // 게시글 타입 (공지글, 모임글, 자유글, 인사글)
     private Long boardCnt;          // 게시글 조회수
